extract compiler to a separate package for classifiers
